backend/requirements.txt
# Backend Python requirements pour Santhium

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Sécurité et authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Cryptographie pour chiffrement des fichiers
cryptography==41.0.7

# Variables d'environnement
python-dotenv==1.0.0

# Validation de données
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# Gestion des fichiers
aiofiles==23.2.1

# Utilitaires
python-dateutil==2.8.2
pytz==2023.3

# Tests (optionnel en dev)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2